Started Changes to get decrypted game binary

+++ Arty-Z7-10/components/ext_sources/u-boot-ectf/include/mesh_crypto.h
@@ -9,4 +9,4 @@ void *safe_malloc(size_t size);
 void safe_free(void* ptr, size_t size);
 int verify_user_can_play(char *username, char* pin, char* gamepath);
 void decrypt_game_file(char *username, char* pin, char* gamepath);
-int crypto_get_game_header(Game *game, char *game_name);
+loff_t crypto_get_game_header(Game *game, char *game_name);
